data_ingestion/collectors.py
"""
Data collectors for various market data sources.
"""

import asyncio
import pandas as pd
from abc import ABC, abstractmethod
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Any

from tvdatafeed import TvDatafeed
from alpaca.data.historical import StockHistoricalDataClient
from alpaca.data.requests import StockBarsRequest
from alpaca.data.timeframe import TimeFrame

from utils.logging import get_logger
from utils.config import get_settings

logger = get_logger(__name__)


class DataCollector(ABC):
    """Abstract base class for data collectors."""
    
    @abstractmethod
    async def get_historical_data(
        self, 
        symbol: str, 
        start_date: datetime, 
        end_date: datetime,
        timeframe: str = "1D"
    ) -> pd.DataFrame:
        """Get historical OHLCV data."""
        pass
    
    @abstractmethod
    async def get_real_time_data(self, symbols: List[str]) -> Dict[str, Any]:
        """Get real-time market data."""
        pass


class AlpacaDataCollector(DataCollector):
    """Alpaca data collector for US equities."""
    
    def __init__(self):
        self.settings = get_settings()
        # TODO: Initialize Alpaca data client
        # self.client = StockHistoricalDataClient(
        #     api_key=self.settings.alpaca_api_key,
        #     secret_key=self.settings.alpaca_secret_key
        # )
        logger.info("Initialized Alpaca data collector")
    
    async def get_historical_data(
        self, 
        symbol: str, 
        start_date: datetime, 
        end_date: datetime,
        timeframe: str = "1Day"
    ) -> pd.DataFrame:
        """
        Get historical data from Alpaca.
        
        Args:
            symbol: Trading symbol
            start_date: Start date
            end_date: End date  
            timeframe: Data timeframe
            
        Returns:
            DataFrame with OHLCV data
            
        TODO: Implement actual Alpaca data retrieval
        """
        logger.info(f"Fetching historical data for {symbol} from Alpaca")
        
        # TODO: Implement actual data fetching
        # request = StockBarsRequest(
        #     symbol_or_symbols=symbol,
        #     timeframe=TimeFrame.Day,
        #     start=start_date,
        #     end=end_date
        # )
        # bars = self.client.get_stock_bars(request)
        
        # Placeholder data
        dates = pd.date_range(start=start_date, end=end_date, freq='D')
        data = {
            'timestamp': dates,
            'open': 100.0,
            'high': 105.0,
            'low': 95.0,
            'close': 102.0,
            'volume': 1000000
        }
        
        return pd.DataFrame(data)
    
    async def get_real_time_data(self, symbols: List[str]) -> Dict[str, Any]:
        """Get real-time quotes from Alpaca."""
        # TODO: Implement real-time data streaming
        return {symbol: {"price": 100.0, "timestamp": datetime.utcnow()} 
                for symbol in symbols}


class TradingViewDataCollector(DataCollector):
    """TradingView data collector."""
    
    def __init__(self):
        self.settings = get_settings()
        # TODO: Initialize TradingView client with credentials
        # self.tv = TvDatafeed(
        #     username=self.settings.tradingview_username,
        #     password=self.settings.tradingview_password
        # )
        logger.info("Initialized TradingView data collector")
    
    async def get_historical_data(
        self, 
        symbol: str, 
        start_date: datetime, 
        end_date: datetime,
        timeframe: str = "1D"
    ) -> pd.DataFrame:
        """
        Get historical data from TradingView.
        
        TODO: Implement TradingView data fetching
        """
        logger.info(f"Fetching historical data for {symbol} from TradingView")
        
        # TODO: Implement actual TradingView data fetching
        # data = self.tv.get_hist(
        #     symbol=symbol,
        #     exchange='NASDAQ',
        #     interval=Interval.in_daily,
        #     n_bars=1000
        # )
        
        # Placeholder data
        dates = pd.date_range(start=start_date, end=end_date, freq='D')
        data = {
            'timestamp': dates,
            'open': 100.0,
            'high': 105.0,
            'low': 95.0,
            'close': 102.0,
            'volume': 1000000
        }
        
        return pd.DataFrame(data)
    
    async def get_real_time_data(self, symbols: List[str]) -> Dict[str, Any]:
        """Get real-time data from TradingView."""
        # TODO: Implement TradingView real-time data
        return {symbol: {"price": 100.0, "timestamp": datetime.utcnow()} 
                for symbol in symbols}


class YFinanceDataCollector(DataCollector):
    """Yahoo Finance data collector."""
    
    def __init__(self):
        logger.info("Initialized Yahoo Finance data collector")
    
    async def get_historical_data(
        self, 
        symbol: str, 
        start_date: datetime, 
        end_date: datetime,
        timeframe: str = "1d"
    ) -> pd.DataFrame:
        """
        Get historical data from Yahoo Finance.
        
        TODO: Implement yfinance data fetching
        """
        logger.info(f"Fetching historical data for {symbol} from Yahoo Finance")
        
        # TODO: Implement yfinance
        # import yfinance as yf
        # ticker = yf.Ticker(symbol)
        # data = ticker.history(start=start_date, end=end_date, interval=timeframe)
        
        # Placeholder data
        dates = pd.date_range(start=start_date, end=end_date, freq='D')
        data = {
            'timestamp': dates,
            'open': 100.0,
            'high': 105.0,
            'low': 95.0,
            'close': 102.0,
            'volume': 1000000
        }
        
        return pd.DataFrame(data)
    
    async def get_real_time_data(self, symbols: List[str]) -> Dict[str, Any]:
        """Get real-time data from Yahoo Finance."""
        # TODO: Implement real-time Yahoo Finance data
        return {symbol: {"price": 100.0, "timestamp": datetime.utcnow()} 
                for symbol in symbols}


class DataCollectorFactory:
    """Factory for creating data collectors."""
    
    _collectors = {
        'alpaca': AlpacaDataCollector,
        'tradingview': TradingViewDataCollector,
        'yfinance': YFinanceDataCollector,
    }
    
    @classmethod
    def create_collector(cls, source: str) -> DataCollector:
        """
        Create a data collector for the specified source.
        
        Args:
            source: Data source name
            
        Returns:
            DataCollector instance
            
        Raises:
            ValueError: If source is not supported
        """
        if source not in cls._collectors:
            raise ValueError(f"Unsupported data source: {source}")
        
        return cls._collectors[source]()
    
    @classmethod
    def list_sources(cls) -> List[str]:
        """List available data sources."""
        return list(cls._collectors.keys()) 