pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "algua"
version = "1.0.0"
description = "Quantitative trading platform with VectorBT Pro, PyTorch, and FastAPI"
authors = [
    {name = "Algua Team", email = "team@algua.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "streamlit>=1.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "pytest>=7.0.0",
    "optuna>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
ml = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "wandb>=0.15.0",
    "scikit-learn>=1.3.0",
]
trading = [
    "vectorbtpro>=1.0.0",
    "alpaca-trade-api>=3.0.0",
    "alpaca-py>=0.9.0",
    "tvdatafeed>=2.0.0",
    "yfinance>=0.2.0",
    "ta-lib>=0.4.0",
]

[project.urls]
homepage = "https://github.com/algua/algua"
repository = "https://github.com/algua/algua"
documentation = "https://algua.readthedocs.io"

[project.scripts]
algua-backtest = "scripts.run_backtest:main"
algua-optimize = "scripts.optimize_strategy:main"
algua-live = "scripts.run_live_trading:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    'tests/',
    'notebooks/',
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*" 